cleaned up a bit, split javascript into parts for each team, changed compilation instructions
