+++ .gitignore
@@ -3,4 +3,5 @@ fastcluster_dm.cpp
 fastcluster_R_dm.cpp
 fastcluster.cpp
 fastcluster.h
-
+.DS_Store
+.vscode/
