+++ home.html
@@ -1,39 +1,45 @@
 <!doctype html>
 <html lang="en-US">
-    <head>
-        <meta charset="utf-8"/>
-        <mate name="viewport" content="width=device-width"/>
-        <title>Data Visualizer</title>
-        <link href="style/generalstyle.css" rel="stylesheet">
-        <link href="style/actionstyle.css" rel="stylesheet">
-        <script src="script/mainscript.js" defer></script>
-    </head>
+  <head>
+    <meta charset="utf-8" />
+    <mate name="viewport" content="width=device-width" />
+    <title>Data Visualizer</title>
+    <link href="style/generalstyle.css" rel="stylesheet" />
+    <link href="style/actionstyle.css" rel="stylesheet" />
+    <script src="src/js/mainscript.js" defer></script>
+  </head>
 
-    <body>
-        <header>
-            <h1>Data Visualizer</h1>
-            <p  id="welcome">Welcome!</p>
-        </header>
-        <main>
-            <div id="introduction_text">
-                <h2>Intruduction</h2>
-                <p id="usecase">Welcome to the Data Visualizer tool. This tool allows you to upload data in CSV format, define functions, and visualize the results. Follow the instructions below to get started:</p>
-                <ol>
-                    <li>Click "Add Data" to upload a CSV file with your data.</li>
-                <li>Type or edit your custom functions in the textarea provided.</li>
-                <li>Choose a functions for visualization.</li>
-                <li>Click "Run" to execute the selected function on your data.</li>
-                <li>Use "Delete All" to clear the function input field.</li>
-                </ol>
-            </div>
-            <div  style="text-align: center;">
-                <button id="start" onclick="location.replace('main.html')" >Get started!</button>
-            </div>
-        </main>
-    </body> 
+  <body>
+    <header>
+      <h1>Data Visualizer</h1>
+      <p id="welcome">Welcome!</p>
+    </header>
+    <main>
+      <div id="introduction_text">
+        <h2>Intruduction</h2>
+        <p id="usecase">
+          Welcome to the Data Visualizer tool. This tool allows you to upload
+          data in CSV format, define functions, and visualize the results.
+          Follow the instructions below to get started:
+        </p>
+        <ol>
+          <li>Click "Add Data" to upload a CSV file with your data.</li>
+          <li>Type or edit your custom functions in the textarea provided.</li>
+          <li>Choose a functions for visualization.</li>
+          <li>Click "Run" to execute the selected function on your data.</li>
+          <li>Use "Delete All" to clear the function input field.</li>
+        </ol>
+      </div>
+      <div style="text-align: center">
+        <button id="start" onclick="location.replace('main.html')">
+          Get started!
+        </button>
+      </div>
+    </main>
+  </body>
 
-    <footer>
-        <h2>Thanks to:</h2>
-        <p id="stuff">...</p>
-    </footer>
+  <footer>
+    <h2>Thanks to:</h2>
+    <p id="stuff">...</p>
+  </footer>
 </html>
