+++ mapstuff.html
@@ -1,23 +1,22 @@
-<!DOCTYPE html>
+<!doctype html>
 <html lang="en-US">
-    <head>
-        <meta charset="utf-8"/>
-        <script src="https://d3js.org/d3.v7.min.js"></script>
-        <mate name="viewport" content="width=device-width"/>
-        <title>Data Visualizer</title>
-        <link href="style/generalstyle.css" rel="stylesheet">
-        <link href="style/actionstyle.css" rel="stylesheet">
-        <script src="script/interact.js" ></script>
+  <head>
+    <meta charset="utf-8" />
+    <script src="https://d3js.org/d3.v7.min.js"></script>
+    <mate name="viewport" content="width=device-width" />
+    <title>Data Visualizer</title>
+    <link href="style/generalstyle.css" rel="stylesheet" />
+    <link href="style/actionstyle.css" rel="stylesheet" />
+    <script src="script/interact.js"></script>
 
-        <!--   Site-spesificiale Style   -->
-    </head>
+    <!--   Site-spesificiale Style   -->
+  </head>
 </html>
 
 <body>
-    
-    <div id="chartContainer"></div>
-    <script src="map.js"></script>
-    <script>
-        initializeMap();
-    </script>
+  <div id="chartContainer"></div>
+  <script src="map.js"></script>
+  <script>
+    initializeMap();
+  </script>
 </body>
