+++ style/generalstyle.css
@@ -1,51 +1,46 @@
-
-body{
-    
-    font-family: "Inter", sans-serif;
-    color: black;
-    background-color: white;
-    margin: 0;
-    padding: 0;
+body {
+  font-family: "Inter", sans-serif;
+  color: black;
+  background-color: white;
+  margin: 0;
+  padding: 0;
 }
-p{
-    text-align: left;
+p {
+  text-align: left;
 }
 
-header{
-    width: 100%;
-    padding: 15px;
-    background-color: #ebaccb; 
+header {
+  width: 100%;
+  padding: 15px;
+  background-color: #ebaccb;
 }
 
-
-main{
-    width: 100%;
-    height: 100%;
-    padding: 15px;
-    background-color: white;
+main {
+  width: 100%;
+  height: 100%;
+  padding: 15px;
+  background-color: white;
 }
 
-footer{
-    width: 100%;
-    min-height: 40px;
-    text-align: left;
-    padding: 5px;
-    background-color: white;
+footer {
+  width: 100%;
+  min-height: 40px;
+  text-align: left;
+  padding: 5px;
+  background-color: white;
 }
 
-h1{
-    text-align: center;
+h1 {
+  text-align: center;
 }
 
-
-button{
-    display: inline-block;
-    width:150px;
-    background-color: #969FE2;
-    margin: 2px;
-    padding: 10px 15px;
-    border-radius: 5px;
-    cursor: pointer;
-    border-color: transparent;
+button {
+  display: inline-block;
+  width: 150px;
+  background-color: #969fe2;
+  margin: 2px;
+  padding: 10px 15px;
+  border-radius: 5px;
+  cursor: pointer;
+  border-color: transparent;
 }
-
